+++ librecraft/src/graphics/transform.cpp
@@ -13,7 +13,7 @@ namespace graphics
 // Constructor/Destructor #################################
 // ########################################################
 
-Transform::Transform(float x, float y, float z)
+Transform::Transform(double x, double y, double z)
 {
     translation = glm::translate(glm::mat4(1.f), glm::vec3(x, y, z));
 
@@ -30,36 +30,36 @@ void Transform::translate(glm::vec3 value) {
     translation = glm::translate(translation, value);
 }
 
-void Transform::translate(float x, float y, float z) {
+void Transform::translate(double x, double y, double z) {
     translation = glm::translate(translation, glm::vec3(x, y, z));
 }
 
-void Transform::setLocation(float x, float y, float z) {
+void Transform::setLocation(double x, double y, double z) {
     translation = glm::translate(glm::mat4(1.f), glm::vec3(x, y, z));
 }
 
-void Transform::setRotationX(float value) {
-    xRotation = glm::rotate(glm::mat4(1.0f), value, glm::vec3(1, 0, 0));
+void Transform::setRotationX(double value) {
+    xRotation = glm::rotate(glm::mat4(1.0f), static_cast<float>(value), glm::vec3(1, 0, 0));
 }
 
-void Transform::setRotationY(float value) {
-    yRotation = glm::rotate(glm::mat4(1.0f), value, glm::vec3(0, 1, 0));
+void Transform::setRotationY(double value) {
+    yRotation = glm::rotate(glm::mat4(1.0f), static_cast<float>(value), glm::vec3(0, 1, 0));
 }
 
-void Transform::setRotationZ(float value) {
-    zRotation = glm::rotate(glm::mat4(1.0f),value, glm::vec3(0, 0, 1));
+void Transform::setRotationZ(double value) {
+    zRotation = glm::rotate(glm::mat4(1.0f), static_cast<float>(value), glm::vec3(0, 0, 1));
 }
 
-void Transform::rotateX(float value) {
-    xRotation = glm::rotate(xRotation, value, glm::vec3(1, 0, 0));
+void Transform::rotateX(double value) {
+    xRotation = glm::rotate(xRotation, static_cast<float>(value), glm::vec3(1, 0, 0));
 }
 
-void Transform::rotateY(float value) {
-    yRotation = glm::rotate(yRotation, value, glm::vec3(0, 1, 0));
+void Transform::rotateY(double value) {
+    yRotation = glm::rotate(yRotation, static_cast<float>(value), glm::vec3(0, 1, 0));
 }
 
-void Transform::rotateZ(float value) {
-    zRotation = glm::rotate(zRotation, value, glm::vec3(0, 0, 1));
+void Transform::rotateZ(double value) {
+    zRotation = glm::rotate(zRotation, static_cast<float>(value), glm::vec3(0, 0, 1));
 }
 
 glm::mat4 Transform::getMatrix() {
