+++ librecraft/src/gui/image.cpp
@@ -9,8 +9,8 @@ namespace gui {
 // Constructor/Destructor #################################
 // ########################################################
 
-Image::Image(int x, int y, int width, int height, const std::string image) {
-	sprite.reset(new graphics::Sprite {x, y, 0, width, height,
+Image::Image(double x, double y, double width, double height, const std::string image) {
+	sprite.reset(new graphics::Sprite{x, y, 0, width, height,
 			graphics::Resources::getInstance().getTexture(image)});
 }
